+++ lp-model-verification/verification-component-understandability-plugin/src/main/java/eu/learnpad/verification/plugin/bpmn/guideline/impl/notationusage/ExplicitStartEndEvents.java
@@ -1,4 +1,4 @@
-package eu.learnpad.verification.plugin.bpmn.guideline.impl;
+package eu.learnpad.verification.plugin.bpmn.guideline.impl.notationusage;
 
 
 import org.eclipse.bpmn2.Definitions;
@@ -9,6 +9,8 @@ import org.eclipse.bpmn2.Process;
 import org.eclipse.bpmn2.RootElement;
 import org.eclipse.bpmn2.StartEvent;
 import org.eclipse.bpmn2.SubProcess;
+
+import eu.learnpad.verification.plugin.bpmn.guideline.impl.abstractGuideline;
 
 public class ExplicitStartEndEvents extends abstractGuideline{
