+++ lp-model-verification/verification-component-understandability-plugin/src/test/java/eu/learnpad/verification/plugin/bpmn/guideline/factory/GuidelinesFactoryTest.java
@@ -24,7 +24,7 @@ public class GuidelinesFactoryTest {
 	@Test
 	public void testGuidelinesFactoryListOfRootElement() {
 		genarateTestforFile("24485.bpmn","16");
-		/*genarateTestforFile("ExplicitStartEndEvents.bpmn","12");
+		genarateTestforFile("ExplicitStartEndEvents.bpmn","12");
 		genarateTestforFile("annidategateway.bpmn","16");
 		genarateTestforFile("SplitAndJoinFlows.bpmn","18");
 		genarateTestforFile("modelloTest.bpmn","18");
@@ -34,7 +34,7 @@ public class GuidelinesFactoryTest {
 		genarateTestforFile("test7.bpmn","13");
 		genarateTestforFile("test8.bpmn","15");
 		genarateTestforFile("MeaningfulGateways.bpmn","20");
-		genarateTestforFile("testorg.bpmn","21");*/
+		genarateTestforFile("testorg.bpmn","21");/**/
 
 	}
